Change result type to long and add warning

+++ src/main/java/pl/rengreen/reversedigits/Digits.java
@@ -2,12 +2,12 @@ package pl.rengreen.reversedigits;
 
 class Digits {
 
-    int reverseDigits(int number){
-        int result=0;
+    long reverseDigits(int number) {
+        long result = 0;
 
-        while (number>0){
-        result=result*10+number%10;
-        number=number/10;
+        while (number > 0) {
+            result = result * 10 + number % 10;
+            number = number / 10;
         }
 
         return result;
